Add PHPStan static analysis at level 6
- phpstan.neon config with Moodle bootstrap and appropriate ignores
  for Moodle globals ($DB, $CFG, $USER, etc.) and framework conventions
- GitHub Action that checks out Moodle 4.5, symlinks the plugin in,
  and runs PHPStan against it
- composer.json with phpstan/phpstan ^2.0 dev dependency

Code fixes to pass level 6:
- client.php: Use strict base64_decode() with is_string() check
- config.php: Cast FORMAT_PLAIN to int for format_text()
- lib.php: Add explicit return true after send_stored_file()

Also adds .gitignore for vendor/ and composer.lock.

+++ config.php
@@ -125,7 +125,7 @@ if ($logs) {
             userdate($log->timecreated),
             s($log->status),
             s(substr($log->commit_sha ?? '', 0, 7)),
-            format_text($log->summary, FORMAT_PLAIN),
+            format_text($log->summary, (int) FORMAT_PLAIN),
         ];
     }
     echo html_writer::table($table);
